double to decimal

+++ AbySalto.Junior/Infrastructure/Entities/OrderArticle.cs
@@ -6,7 +6,7 @@
         public int OrderId { get; set; }
         public int ArticleId { get; set; }
         public int Quantity { get; set; }
-        public double OrderPrice { get; set; }
+        public decimal OrderPrice { get; set; }
 
         public Order? Order { get; set; }
         public Article? Article { get; set; }
