+++ AbySalto.Junior/Models/OrderArticleDto.cs
@@ -8,6 +8,6 @@ namespace AbySalto.Junior.Models
         public int OrderId { get; set; }
         public int ArticleId { get; set; }
         public int Quantity { get; set; }
-        public double OrderPrice { get; set; }
+        public decimal OrderPrice { get; set; }
     }
 }
